Admin and Author
Allows admin users to write more than one message

Now locking instead of archiving

Auto archive set to a week

Multiple message refactor

Edit Prevention

+++ commands/ask.js
@@ -1,28 +1,42 @@
-const { SlashCommandBuilder, PermissionFlagsBits} = require('discord.js'); 
+const { SlashCommandBuilder, PermissionFlagsBits, Events, ThreadAutoArchiveDuration} = require('discord.js'); 
 
 module.exports = {
     data: new SlashCommandBuilder()
     .setName('ask')
     .setDescription('Starts a AskOuija Question')
     .setDefaultMemberPermissions(PermissionFlagsBits.Administrator)
+    .setDMPermission(false)
     .addStringOption(option => option
         .setName('question')
-        .setDescription('A question')
+        .setDescription('What to ask?')
         .setRequired(true)),
     async execute(interaction) {
+        if(interaction.channel.isThread())
+            return
         const question = interaction.options.getString('question');
         await interaction.deferReply();
 
         const thread = await interaction.channel.threads.create({
             name: question,
-            reason: 'A new AskOuija question'
+            reason: 'A new AskOuija question',
+            autoArchiveDuration: ThreadAutoArchiveDuration.OneWeek
         });
+
         if (thread.joinable) await thread.join();
 
+        thread.client.on(Events.MessageUpdate, (oldMessage, newMessage) => {
+            if(!newMessage.member.permissions.has(PermissionFlagsBits.Administrator))
+                collector.collected.set(newMessage.id, oldMessage);
+        })
+
+        thread.client.on(Events.MessageDelete, (message) => {
+            collector.authors.delete(message.author.id)
+        })
+
         let collector = thread.createMessageCollector({dispose: true});
-        collector.authors = {};
+        collector.authors = new Map();
         collector.on('collect', message => {
-            if(message.author.id in collector.authors){
+            if(message.author.id in collector.authors && !message.member.permissions.has(PermissionFlagsBits.Administrator)){
                 collector.collected.delete(message.id);
                 message.delete();
             } else if(message.content.toUpperCase() === 'GOODBYE'){
@@ -32,18 +46,18 @@ module.exports = {
                 collector.collected.delete(message.id);
                 message.delete();
             } else {
-                collector.authors[message.author.id] = true;
+                collector.authors.set(message.author.id, true);
             }
         })
         collector.on('end', async collected => {
             const answer = collected.map(value => value.content).join('').toUpperCase();
             let response = `\`\`\`yaml\nQuestion: ${question} \nAnswer: ${answer} \n\`\`\``;
             if(question.includes('_')){
-                response = `\`\`\`yaml\n${question.replace('_', answer)}\n\`\`\``;
+                response = `\`\`\`yaml\n${question.replace(/_*/i, answer)}\n\`\`\``;
+                await thread.setName(response)
             }
             await thread.send(response);
             await thread.setLocked(true);
-            await thread.setArchived(true);
             await interaction.editReply(response)
         })
     }
